moving isChoreOverdue function to viewModel and applying to prevChores filtering
